requirements.txt
Flask==2.3.3
Pillow==10.0.0
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.0
tqdm==4.66.0
